Refactoring: moving KEML relevant parts into shared keml package
