+++ src/app/features/editor/features/editor/editor.component.ts
@@ -13,9 +13,9 @@ import { MsgComponent } from '@app/shared/keml/components/msg/msg.component';
 import { PreknowledgeComponent } from '@app/shared/keml/components/preknowledge/preknowledge.component';
 import { ConversationPartnerComponent } from '@app/shared/keml/components/cp/conversation-partner.component';
 import { AuthorComponent } from '@app/shared/keml/components/author/author.component';
-import { TextAreaSvgComponent } from '@app/core/features/svg-base-components/text-area-svg/text-area-svg.component';
-import { DatabaseSvgComponent } from '@app/core/features/svg-base-components/database-svg/database-svg.component';
-import { PersonSvgComponent } from '@app/core/features/svg-base-components/person-svg/person-svg.component';
+import { TextAreaSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/text-area-svg/text-area-svg.component';
+import { DatabaseSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/database-svg/database-svg.component';
+import { PersonSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/person-svg/person-svg.component';
 import {ConversationPickerComponent} from '@app/features/editor/fromLLM/components/conversation-picker/conversation-picker.component';
 import {SimulationService} from "@app/features/simulator/services/simulation.service";
 import {ArrowMarkersComponent} from "@app/shared/keml/components/helper/arrow-markers/arrow-markers.component";
