+++ src/app/features/simulator/components/simulation-input-details/simulation-input-details.component.ts
@@ -6,7 +6,7 @@ import {SimulationInputs} from "@app/features/simulator/models/simulation-inputs
 import {NgForOf} from "@angular/common";
 import {ConversationPartner} from "@app/shared/keml/models/core/conversation-partner";
 import {TrustComputator} from "@app/features/simulator/utils/trust-computator";
-import {TrustSliderComponent} from "@app/shared/trust-slider/trust-slider.component";
+import {TrustSliderComponent} from "@app/shared/keml/components/helper/trust-slider/trust-slider.component";
 
 @Component({
     selector: 'simulation-input-details',
