+++ src/app/shared/keml/components/helper/msg-overview/msg-overview.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input} from '@angular/core';
 import {Message} from "@app/shared/keml/models/core/msg-info";
-import { TextAreaSvgComponent } from '@app/core/features/svg-base-components/text-area-svg/text-area-svg.component';
-import { PersonSvgComponent } from '@app/core/features/svg-base-components/person-svg/person-svg.component';
+import { TextAreaSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/text-area-svg/text-area-svg.component';
+import { PersonSvgComponent } from '@app/shared/keml/components/helper/svg-base-components/person-svg/person-svg.component';
 import { MsgInnerComponent } from '../msg-inner/msg-inner.component';
 
 @Component({
