+++ src/app/shared/keml/components/helper/svg-base-components/database-svg/database-svg.component.ts
