+++ src/app/shared/keml/components/helper/svg-base-components/person-svg/person-svg.component.spec.ts
