+++ src/app/shared/keml/components/helper/svg-base-components/text-area-svg/text-area-svg.component.spec.ts
