+++ src/app/shared/keml/components/helper/trust-slider/trust-slider.component.ts
