+++ src/app/shared/keml/utils/trust-colorer.spec.ts
