+++ src/app/shared/keml/utils/trust-colorer.ts
